feat: Add pre_restart_hook
This script is expected to be hooked into valheim-server-docker's
`PRE_RESTART_HOOK` to run before the valheim-updater restarts
the server after applying any updates to it. It sends a message
stating that the server will restart in 1 minute to a Discord
webhook, then sleeps for 1 minute before ending. At which point,
the server will actually restart