feat: Add script to announce when a user connects
This script is expected to be triggered when the log message
for a user connecting to the server appears. It parses the log
message to find the user's character's name, then sends a message
stating that the user has connected, to a Discord webhook. It also
writes that name and the ZDOID to a line in a mapping file that
can be referenced while that user is still connected to get their
character's name from the ZDOID.

Example log message expected to trigger this script:
```
Got character ZDOID from <CHARACTER_NAME> : 2605381299:1
```