feat: Add script to announce when user disconnects
This script is expected to be triggered when log messages
related to a user disconnecting from the server appear. It
sends a message saying that the user has disconnected to a
Discord webhook. It also removes the user:ZDOID mapping from
the mapping file, as it is no longer useful, as the next time
the user logs in, they will have a different ZDOID. If the
mapping doesn't exist, then it exits early without sending
the message.

These messages are not formatted as nicely as the connection
ones are. They don't provide the character's name and only
provide the ZDOID in log messages for destroying resources owned
by the ZDOID after the user disconnects. There are several of
those log messages with no easy way to only trigger on one of
them. So, this script has to take some unusual steps to ensure
it only fully reacts to one of these events and sends the message.

Since the script can end up running 2 or more times in quick
succession, it will check if a lock file has been created for this
event, kept unique per user logout event by using the ZDOID. If
the lock file exists, it exits early. If the lock file doesn't
exist, it creates it, then performs the rest of the functionality
mentioned above, then removes the lock file.

I found that sometimes the scripts run in such quick succession
that the lock file wasn't even created yet by the time other
instances do the check for it, so multiple instances were still
reacting to the event and sending duplicate messages. So, I added
a short, random sleep (a few MS at most), to spread the lock
file check out a bit and ensure one of them is able to actually
acquire the lock.

Example log messages for a user disconnecting:
```
RPC_Disconnect
Destroying abandoned non persistent zdo 2605381299:7 owner 2605381299
Destroying abandoned non persistent zdo 2605381299:10 owner 2605381299
Destroying abandoned non persistent zdo 2605381299:13 owner 2605381299
```

The RPC_Disconnect is the actual disconnect event. Then it starts
destroying ZDO's that were tied to that ZDOID. In this case, those 3
log lines starting wtih "Destroying abandoned..." are what we expect
to trigger this script. So there will be 3 triggers for this one
disconnect event.